Make runModal return NSInteger
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@36696 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormSetNameController.h
@@ -17,7 +17,7 @@
   NSButton *okButton;
   NSButton *cancelButton;
 }
-- (int)runAsModal;
+- (NSInteger)runAsModal;
 
 - (NSTextField *) textField;
 - (void) cancelHit: (id)sender;
